For issue #13: Tested notifications GET on the server and /o2m/notifications client view.
All features working properly but notifications only have titles, not links or anythin else. But this proof of concept shows that it should not be dificult to add links in a similar way so that notifications provide a more useful functionality as designed.

Also, they do not have the 'seen' property yet.