Problems with obj_creator. Now post notifications are sent correctly (to the owner included for the time being).
So now notifications have a creator, extra information can be given to the receiever. The creator must be a friend.

When a comment is made on someone's post, a notification is sent to them. At the moment, this is identical to the notification sent when a comment is made on a comment. This can be changed easily. The 'obj' in this case is the id of the Link added so that some template in the future can provide a link for the user to click on to get to the 'obj'. This is the theory for every notification - 'obj' should be reachable somehow so there should be enough information in the notification together with its type to do that.