If a friend is browsed to using non_friend, add them
When someone is not a friend, there still needs to be a way to access their page. This can be done with the non_friend url. When a user does this, a Friend row is added to the Friend model and the password for accessing their IP and port is set to 'NOTFRIENDS' to be extra clear. Later this could be a flag. If people want to allow friends straight away, they could just add a User to their server withe same password (which would get updated imediately).