BIG CHANGES for setuptools
OK so, I've tested this for a setup on a Mac. The o2m.settings are using *nix-specific directories so it's time to use distutils resource capabilities. But the fact that it just works on a mac is promising. Installation is still on the command line. The next step will be to find out how setuptools can automatically choose a directory for resources that are created during the running of the program. That way, if it's installed on Windows, the users' files will be in a folder that they are familiar with.