Slightly better cross-platform support
I say slightly because in an attempt to fix issue #16, all files are now placed in the site-packages directory of the Python installation being used by the package manager. This is not technically very good because really, there should just be code in there, when really you also get users' social files in there.