Cache control on content
Now content will become cached for the next minute after which it will be asked for again. This should be extended to infinity and then a flag can be used on each Content to signify whether it should be requested with the "no-cache" header. Then when content is edited on a friend's machine, the flag is flipped to "out of date".