Upload files to basic_server /content/[id] endpoint
Files can now be uploaded to the server endpoint. This will be important for making a more advanced post mechanism that can accept all types of file. The basic client uses AJAX to post a form containing a file picker directly to the server.

Some code needs removing as it is no longer needed.

basic_client views.py needs a major refactoring