Successful instalation onto Mac
Another test was done after some changes to reposition the O2M user files to the platform-independant user directory. This required the specfic command 'python setup.py install --user'. The next step is to create an MSI file for windows and an .app for mac. This should make it very easy to distribute o2m.

Something to think about is how to distribute updates. Major versions should have MSIs and Apps but minor versions will have to be installed with setup.py I think.