Closes #24 - Very robust password updating
The password updating has been made even more robust to errors that may occur on the Client. You can pretty much spam the refresh button on the browser and the worst that will happen is that some of the requests for Friends' timelines will return 401 (but the responses are lost in the midst of anarchy anyway).

This issue highlights that maybe passwords should just stay the same and traditional security techniques should be used (because this feature will cause many many headaches for future implementations - especially because it invalidates layers such as the cache, which some routers will not treat nicely!) Certificates and SSL would be much better.